+++ crates/napi/types/config.d.ts
