+++ crates/napi/types/node-types.d.ts
